+++ gpwgen/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "gpwgen"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "gpwgen"
+path = "src/lib.rs"
+
+[[bin]]
+name = "gpwgen"
+path = "src/main.rs"
+
+[dependencies]
+clap = {version = "*", features = ["derive"]}
+geo = "*"
+hextree = "*"
+rayon = "*"
