+++ gpwgen/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod args;
+pub mod error;
+pub mod generate;
+pub mod gpwascii;
