+++ gpws/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "gpws"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bincode = "*"
+clap = {version = "*", features = ["derive"]}
+hextree = {version = "*", features = ["serde-support"]}
+hyper = {version = "*", features = ["server", "http1", "full"]}
+tokio = {version = "*", features = ["full"]}
+
+# bytes = "1.3.0"
+# geo = "0.23.0"
+# hextree = {git = "https://github.com/JayKickliter/HexTree.git", rev = "ca7d1dac68fc01a1b0bfe48f6481cb23d9d23b28", features = ["serde-support"]}
+# http-body-util = "0.1.0-rc.2"
+# num-traits = "0.2.15"
+# rayon = "1.6.1"
